+++ include/asm-generic/bitops/sched.h
@@ -17,12 +17,12 @@ static inline int sched_find_first_bit(const unsigned long *b)
 	return __ffs(b[1]) + 64;
 #elif BITS_PER_LONG == 32
 	if (b[0])
-		return __ffs(b[0]);
+		return __ffs((long)b[0]);
 	if (b[1])
-		return __ffs(b[1]) + 32;
+		return __ffs((long)b[1]) + 32;
 	if (b[2])
-		return __ffs(b[2]) + 64;
-	return __ffs(b[3]) + 96;
+		return __ffs((long)b[2]) + 64;
+	return __ffs((long)b[3]) + 96;
 #else
 #error BITS_PER_LONG not defined
 #endif
